Mark preStop signal so that we can fail readiness checks
Apparently readiness checks are still being sent by k8s after
a preStop signal, and the load balancer still routes to the
pods that are in terminating state.

This is an experimental change to create a marker file so that
we can fail readiness (but not liveness!!!) checks from the
Django app.